requirements/base.txt
django==4.0.6
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
djangorestframework-simplejwt==5.3.0
django-modeltranslation
Pillow